+++ src/extract/selector.rs
@@ -1,4 +1,5 @@
 mod cache;
+mod compiled;
 mod element;
 mod list;
 mod regex;
@@ -6,6 +7,7 @@ mod regex;
 pub(crate) type SharedDocument = std::sync::Arc<std::sync::Mutex<scraper::Html>>;
 
 pub(crate) use cache::get_selector;
+pub use compiled::CssSelector;
 pub use element::Element;
 pub use list::ElementList;
 pub(crate) use regex::re_matches;